Lab 01/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.54.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
]
